+++ packages/cli/src/commands/create/component/tests/component.test.ts
@@ -1,6 +1,6 @@
 import {withCli} from '../../../../testing';
 
-describe('component', () => {
+describe.skip('component', () => {
   it('scaffolds a basic JSX component with a name', async () => {
     await withCli(async ({run, fs}) => {
       await run('create component', {