+++ packages/hydrogen/src/components/Router/index.tsx
@@ -1,0 +1,5 @@
+export {
+  Router,
+  useRouter,
+  useLocation,
+} from '../../foundation/Router/Router.client';