+++ packages/hydrogen/src/foundation/useShop/use-shop.tsx
@@ -1,18 +1,11 @@
-import {useContext} from 'react';
 import {ShopifyContext} from '../ShopifyProvider';
-import {useServerRequest} from '../ServerRequestProvider';
-// @ts-ignore
-import {META_ENV_SSR} from '../../utilities/meta-env-ssr';
-
-// let contextValue: ShopifyContextValue | null = null;
+import {useEnvContext} from '../ssr-interop';
 
 /**
  * The `useShop` hook provides access to values within `shopify.config.js`. It must be a descendent of a `ShopifyProvider` component.
  */
 export function useShop() {
-  const config = META_ENV_SSR
-    ? useServerRequest().ctx.shopifyConfig
-    : useContext(ShopifyContext);
+  const config = useEnvContext((req) => req.ctx.shopifyConfig, ShopifyContext);
 
   if (!config) {
     throw new Error('No Shopify Context found');