+++ packages/hydrogen/src/framework/Hydration/rsc.ts
@@ -8,6 +8,7 @@ import {
   createFromReadableStream,
   // @ts-ignore
 } from '@shopify/hydrogen/vendor/react-server-dom-vite';
+import {RSC_PATHNAME} from '../../constants';
 
 declare global {
   // eslint-disable-next-line no-var
@@ -80,17 +81,16 @@ export function useServerResponse(state: any) {
 
     // Request a new flight response.
     response = createFromFetch(
-      fetch('/react?state=' + encodeURIComponent(key))
+      fetch(`${RSC_PATHNAME}?state=` + encodeURIComponent(key))
     );
   }
 
+  cache.clear();
   cache.set(key, response);
   return response;
 }
 
 export function useRefresh() {
   const refreshCache = unstable_useCacheRefresh();
-  return function refresh(key: string, seededResponse: any) {
-    refreshCache(createResponseCache, new Map([[key, seededResponse]]));
-  };
+  refreshCache();
 }