+++ packages/hydrogen/src/framework/plugin.ts
@@ -4,6 +4,7 @@ import type {Plugin} from 'vite';
 import hydrogenMiddleware from './plugins/vite-plugin-hydrogen-middleware';
 // @ts-ignore
 import rsc from '@shopify/hydrogen/vendor/react-server-dom-vite/plugin';
+import ssrInterop from './plugins/vite-plugin-ssr-interop';
 import purgeQueryCache from './plugins/vite-plugin-purge-query-cache';
 import inspect from 'vite-plugin-inspect';
 import react from '@vitejs/plugin-react';
@@ -19,6 +20,7 @@ export default (
     hydrogenConfig(),
     hydrogenMiddleware(shopifyConfig, pluginOptions),
     react(),
+    ssrInterop(),
     rsc({
       clientComponentPaths: [
         path.join(